+++ CHANGELOG.md
@@ -205,6 +205,7 @@
 - [10550](https://github.com/vegaprotocol/vega/issues/10550) - Update network parameters with default values.
 - [10612](https://github.com/vegaprotocol/vega/issues/10612) - Convert all assets to be associated to the configured Ethereum chain.
 - [10624](https://github.com/vegaprotocol/vega/issues/10624) - Ensure chain event are not duplicated when chain identifier is missing.
+- [10623](https://github.com/vegaprotocol/vega/issues/10623) - Introduce secondary Ethereum client
 
 ### 🐛 Fixes